+++ src/applications/editors/index.js
@@ -1,6 +1,6 @@
 import CurrenciesEditor from "./currencies-editor/currencies-editor.js";
 import ItemFiltersEditor from "./item-filters-editor/item-filters-editor.js";
-import ItemSimilaritiesEditor from "./item-similarities-editor/item-similarities-editor.js";
+import StringListEditor from "./string-list-editor/string-list-editor.js";
 import PriceModifiersEditor from "./price-modifiers-editor/price-modifiers-editor.js";
 import PricePresetEditor from "./price-preset-editor/price-preset-editor.js";
 import UnstackableItemTypesEditor from "./unstackable-item-types-editor/unstackable-item-types-editor.js";
@@ -10,7 +10,8 @@ import StylesEditor from "./styles-editor/styles-editor.js";
 export default {
   "currencies": CurrenciesEditor,
   "item-filters": ItemFiltersEditor,
-  "item-similarities": ItemSimilaritiesEditor,
+  "item-similarities": StringListEditor,
+  "item-categories": StringListEditor,
   "styles": StylesEditor,
   "vault-styles": VaultStylesEditor,
   "price-modifiers": PriceModifiersEditor,