+++ src/applications/editors/item-filters-editor/item-filters-editor.js
@@ -2,7 +2,7 @@ import ItemFiltersShell from './item-filters-editor.svelte';
 import { SvelteApplication } from '@typhonjs-fvtt/runtime/svelte/application';
 
 export default class ItemFiltersEditor extends SvelteApplication {
-  
+
   constructor(itemFilters, options) {
     super({
       title: "ITEM-PILES.Applications.FilterEditor.Title",
@@ -17,16 +17,15 @@ export default class ItemFiltersEditor extends SvelteApplication {
       ...options
     });
   }
-  
+
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      title: game.i18n.localize("ITEM-PILES.Applications.FilterEditor.Title"),
       width: 400,
       height: "auto",
       classes: ["item-piles-app"]
     })
   }
-  
+
   static async show(itemFilters, options = {}) {
     return new Promise((resolve) => {
       options.resolve = resolve;