+++ src/applications/editors/string-list-editor/string-list-editor.js
@@ -1,35 +1,34 @@
-import ItemSimilaritiesShell from './item-similarities-editor.svelte';
+import StringListEditorShell from './string-list-editor.svelte';
 import { SvelteApplication } from '@typhonjs-fvtt/runtime/svelte/application';
 
-export default class ItemSimilaritiesEditor extends SvelteApplication {
-  
-  constructor(itemSimilarities, options) {
+export default class StringListEditor extends SvelteApplication {
+
+  constructor(stringList, options) {
     super({
       svelte: {
-        class: ItemSimilaritiesShell,
+        class: StringListEditorShell,
         target: document.body,
         props: {
-          itemSimilarities
+          stringList
         }
       },
       close: () => this.options.resolve(null),
       ...options
     });
   }
-  
+
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      title: game.i18n.localize("ITEM-PILES.Applications.SimilaritiesEditor.Title"),
       width: 400,
       height: "auto",
       classes: ["item-piles-app"]
     })
   }
-  
-  static async show(itemSimilarities, options = {}) {
+
+  static async show(stringList, options = {}) {
     return new Promise(resolve => {
       options.resolve = resolve;
-      return new this(itemSimilarities, options).render(true, { focus: true });
+      return new this(stringList, options).render(true, { focus: true });
     });
   }
 }