+++ src/applications/item-editor/item-editor.js
@@ -3,10 +3,10 @@ import { SvelteApplication } from '@typhonjs-fvtt/runtime/svelte/application';
 import { getActiveApps } from '../../helpers/helpers';
 
 export default class ItemEditor extends SvelteApplication {
-  
+
   constructor(item = false, options) {
     super({
-      id: `item-pile-item-editor-${item.id}`,
+      id: `item-pile-item-editor-${item.id}-${randomID()}`,
       title: game.i18n.format("ITEM-PILES.Applications.ItemEditor.Title", { item_name: item.name }),
       svelte: {
         class: ItemEditorShell,
@@ -19,7 +19,7 @@ export default class ItemEditor extends SvelteApplication {
       ...options
     });
   }
-  
+
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
       width: 500,
@@ -27,11 +27,11 @@ export default class ItemEditor extends SvelteApplication {
       classes: ["item-piles-app"]
     })
   }
-  
+
   static getActiveApp(id) {
     return getActiveApps(`item-pile-item-editor-${id}`, true);
   }
-  
+
   static async show(item = false, options = {}, dialogData = {}) {
     const app = this.getActiveApp(item.id);
     if (app) return app.render(false, { focus: true });